Module mina_network::processor::naive_transition_frontier

A naive implementation of the TransitionFrontier

Structs

DummyAccount

Dummy account type for berkeley net

DummyHasher

Dummy merkle hasher for berkeley net

NaiveTransitionFrontier

Struct that represents a naive implementation of the TransitionFrontier

ProtocolStateWithSparseMerkleLedger

protocol state + sparse merkle ledger

Type Definitions

MerkleProofBerkeleyNet

Merkle proof on berkeley net TODO: Replace DummyAccount, DummyHasher with Account, MinaPoseidonMerkleHasher respectively after we can deserialize Account from graphql API response note that there are some new changes to the account hashing algorithm and the ledger hashes have been changed as well. Those have to be fixed accordingly before switching to the real Account

MerkleProofMainnet

Merkle proof on mainnet